PurchasesRecyclerViewAdapter.kt strings

+++ app/src/main/java/com/dezzomorf/financulator/viewmodel/DataBaseViewModel.kt
@@ -124,35 +124,43 @@ open class DataBaseViewModel @Inject constructor(
 
     fun deletePurchasesByCoin(coin: Coin, userId: String) {
         deletePurchasesState.postValue(UiState.Loading)
-        dataBase.collection("users")
-            .document(userId)
-            .collection("purchases")
-            .whereEqualTo("coinId", coin.id)
-            .get()
-            .addOnCompleteListener { task ->
-                if (task.isSuccessful) {
-                    for (document in task.result) {
-                        document.reference.delete()
-                            .addOnSuccessListener {
-                                sharedPreferencesManager.setPurchases(userId, emptyList())
-                                deletePurchasesState.postValue(UiState.Success(Unit))
-                            }
-                            .addOnFailureListener {
-                                deletePurchasesState.postValue(
-                                    UiState.Error(
-                                        Exception(task.exception)
+        if (networkConnectionManager.isConnected.value) {
+            dataBase.collection("users")
+                .document(userId)
+                .collection("purchases")
+                .whereEqualTo("coinId", coin.id)
+                .get()
+                .addOnCompleteListener { task ->
+                    if (task.isSuccessful) {
+                        for (document in task.result) {
+                            document.reference.delete()
+                                .addOnSuccessListener {
+                                    sharedPreferencesManager.setPurchases(userId, emptyList())
+                                    deletePurchasesState.postValue(UiState.Success(Unit))
+                                }
+                                .addOnFailureListener {
+                                    deletePurchasesState.postValue(
+                                        UiState.Error(
+                                            Exception(task.exception)
+                                        )
                                     )
-                                )
-                            }
-                    }
-                } else {
-                    deletePurchasesState.postValue(
-                        UiState.Error(
-                            Exception(task.exception)
+                                }
+                        }
+                    } else {
+                        deletePurchasesState.postValue(
+                            UiState.Error(
+                                Exception(task.exception)
+                            )
                         )
-                    )
+                    }
                 }
-            }
+        } else {
+            deletePurchasesState.postValue(
+                UiState.Error(
+                    Exception()
+                )
+            )
+        }
     }
 
     fun deletePurchase(purchaseId: String, uid: String) {
